Splitting entitycreator.cs
