+++ Assets/Scripts/MonoBehaviour/EntityCreator/ArcEntityCreator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79ee156634405c645a1c89c05aed2b92
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
