+++ Assets/Scripts/MonoBehaviour/EntityCreator/HoldEntityCreator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c88ea1e2b7ad50f47aa6220a224d4ee9
+guid: bd5714a800870224ba23fe4a7e32b93b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
